backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "appwrite==11.0.0",
    "email-validator>=2.3.0",
    "fastapi>=0.116.1",
    "langchain>=0.3.26",
    "langchain-openai>=0.3.28",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
